Add files via upload
Robin Changes

Specials

Elwind 1 - Decreased BKB on spike hit to 10 from 40
(Elwind is a very strong move in multiple ways- as a recovery option due to its high range of variability in timings, as an out of shield option due to its low startup and air actionable state, and as an anti-juggling tool due to the range of the fired projectile, but the strongest use case by far is Elwind's potential as an offstage KO option. At 40 base knockback, the move is on par with Marth's down air in terms of power, but lacking its drawbacks apart from its status as a resource based attack. This power combined with the attack speed and lift gained upon use makes for a very safely noncommital option for securing stocks very early. Lowering the base knockback as proposed brings the move more in line with a Falco down air, being weak at earlier percents but increasing in potency as percents increase.)

Smash Attacks

Down Smash - Grounded hit KBG decreased to 80 from 100