Fox c16-c23 UI
Includes UI for slots c16-c23
chr_11 and chr_13 have a Hazard symbol, being the one used on stages on the top of each render. This is used to warn that (at least I hope it's visible enough) using said skins past slot 16 will not allow you to save replays, but are completely functional otherwise.